requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
awsebcli==3.18.1
Babel==2.8.0
beautifulsoup4==4.9.1
boto3==1.16.6
botocore==1.19.6
bs4==0.0.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
Django==3.1.4
django-cors-headers==3.4.0
django-filter==2.3.0
django-localflavor==3.0.1
django-model-utils==4.0.0
django-phonenumber-field==4.0.0
django-redis-cache==2.1.1
django-rest-framework==0.1.0
django-reverse-admin==2.9.3
django-storages==1.10.1
djangorestframework==3.11.1
docutils==0.15.2
essential-generators==0.9.2
future==0.16.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
iniconfig==1.1.1
jmespath==0.10.0
mock==4.0.3
msgpack==1.0.0
msgpack-python==0.5.6
names==0.3.0
packaging==20.8
pathspec==0.5.9
phonenumbers==8.12.7
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypiwin32==223
pytest==6.2.1
pytest-asyncio==0.14.0
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-stdnum==1.13
pytz==2020.1
pywin32==300
PyYAML==5.3.1
redis==2.10.6
requests==2.24.0
s3transfer==0.3.3
semantic-version==2.5.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.55.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.48.2
twilio==6.45.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
wcwidth==0.1.9
zope.interface==5.1.0
